ipmi: oem_1s_handler: Replace if-statements with sensor_config lookup
Summary:
This code is hard-coding the I2C bus used by the voltage regulators to
i2c5 and using if-statements to map from component to i2c address. This
introduces a lookup table to map to a sensor that includes the same
information, since the sensor table should be the definitive version of
this information.

Test Plan:
Verifying in QEMU that the voltage regulators get queried normally.

Reviewers: goldenbug garnermic

Subscribers:

Tasks:

Tags: